refactor(core): split account tabs to individual pages
