+++ core/app/[locale]/(default)/account/(tabs)/_components/modal.tsx
