+++ core/app/[locale]/(default)/account/(tabs)/addresses/_actions/delete-address.ts
@@ -6,7 +6,7 @@ import { getSessionCustomerId } from '~/auth';
 import { client } from '~/client';
 import { graphql } from '~/client/graphql';
 
-import { State } from './submit-customer-change-password-form';
+import { State } from '../../settings/_actions/submit-customer-change-password-form';
 
 const DeleteCustomerAddressMutation = graphql(`
   mutation DeleteCustomerAddressMutation(
