+++ core/app/[locale]/(default)/account/(tabs)/addresses/_components/customer-new-address.tsx
@@ -10,7 +10,7 @@ import { AddAddress as AddAddressForm } from './add-address';
 
 const CustomerNewAdressQuery = graphql(
   `
-    query CustomerNewAddress(
+    query CustomerNewAdressQuery(
       $countryCode: String
       $shippingFilters: FormFieldFiltersInput
       $shippingSorting: FormFieldSortInput
