+++ core/app/[locale]/(default)/account/(tabs)/settings/_actions/submit-customer-change-password-form.ts
