+++ core/app/[locale]/(default)/account/(tabs)/settings/_components/change-password-form.tsx
