+++ core/app/[locale]/(default)/account/page.tsx
@@ -1,4 +1,4 @@
-import { BookUser, Eye, Gift, Mail, Package, Settings } from 'lucide-react';
+import { BookUser, Settings } from 'lucide-react';
 import { getTranslations } from 'next-intl/server';
 import { ReactNode } from 'react';
 
@@ -41,21 +41,21 @@ export default async function Account({ params: { locale } }: Props) {
       <h1 className="my-8 text-4xl font-black lg:my-8 lg:text-5xl">{t('heading')}</h1>
 
       <div className="mb-14 grid gap-6 md:grid-cols-2 lg:grid-cols-3">
-        <AccountItem href="/account/orders" title={t('orders')}>
+        {/* <AccountItem href="/account/orders" title={t('orders')}>
           <Package className="me-8" size={48} strokeWidth={1.5} />
         </AccountItem>
         <AccountItem href="/account/messages" title={t('messages')}>
           <Mail className="me-8" size={48} strokeWidth={1.5} />
-        </AccountItem>
+        </AccountItem> */}
         <AccountItem href="/account/addresses" title={t('addresses')}>
           <BookUser className="me-8" size={48} strokeWidth={1.5} />
         </AccountItem>
-        <AccountItem href="/account/wishlists" title={t('wishlists')}>
+        {/* <AccountItem href="/account/wishlists" title={t('wishlists')}>
           <Gift className="me-8" size={48} strokeWidth={1.5} />
         </AccountItem>
         <AccountItem href="/account/recently-viewed" title={t('recentlyViewed')}>
           <Eye className="me-8" size={48} strokeWidth={1.5} />
-        </AccountItem>
+        </AccountItem> */}
         <AccountItem href="/account/settings" title={t('settings')}>
           <Settings className="me-8" size={48} strokeWidth={1.5} />
         </AccountItem>
