+++ core/app/[locale]/(default)/login/_components/login-form.tsx
@@ -17,7 +17,7 @@ import {
 } from '~/components/ui/form';
 import { Message } from '~/components/ui/message';
 
-import { useAccountStatusContext } from '../../account/[tab]/_components/account-status-provider';
+import { useAccountStatusContext } from '../../account/(tabs)/_components/account-status-provider';
 import { submitLoginForm } from '../_actions/submit-login-form';
 
 const SubmitButton = () => {
