+++ core/app/[locale]/(default)/login/layout.tsx
@@ -1,6 +1,6 @@
 import { PropsWithChildren } from 'react';
 
-import { AccountStatusProvider } from '../account/[tab]/_components/account-status-provider';
+import { AccountStatusProvider } from '../account/(tabs)/_components/account-status-provider';
 
 export default function LoginLayout({ children }: PropsWithChildren) {
   return <AccountStatusProvider isPermanentBanner={true}>{children}</AccountStatusProvider>;
