+++ core/tests/ui/desktop/e2e/account.spec.ts
@@ -33,32 +33,32 @@ test('My Account tabs are displayed and clickable', async ({ page }) => {
   await loginWithUserAccount(page, testUserEmail, testUserPassword);
 
   await expect(page).toHaveURL('/account/');
-  await expect(page.getByRole('link', { name: 'Orders' })).toBeVisible();
-  await expect(page.getByRole('link', { name: 'Messages' })).toBeVisible();
+  // await expect(page.getByRole('link', { name: 'Orders' })).toBeVisible();
+  // await expect(page.getByRole('link', { name: 'Messages' })).toBeVisible();
   await expect(page.getByRole('link', { name: 'Addresses' })).toBeVisible();
-  await expect(page.getByRole('link', { name: 'Wish lists' })).toBeVisible();
-  await expect(page.getByRole('link', { name: 'Recently viewed' })).toBeVisible();
+  // await expect(page.getByRole('link', { name: 'Wish lists' })).toBeVisible();
+  // await expect(page.getByRole('link', { name: 'Recently viewed' })).toBeVisible();
   await expect(page.getByRole('link', { name: 'Account settings' })).toBeVisible();
 
-  await page.getByRole('link', { name: 'Orders' }).click();
-  await expect(page).toHaveURL('/account/orders/');
-  await expect(page.getByRole('heading', { name: 'Orders' })).toBeVisible();
+  // await page.getByRole('link', { name: 'Orders' }).click();
+  // await expect(page).toHaveURL('/account/orders/');
+  // await expect(page.getByRole('heading', { name: 'Orders' })).toBeVisible();
 
-  await page.getByRole('link', { name: 'Messages' }).click();
-  await expect(page).toHaveURL('/account/messages/');
-  await expect(page.getByRole('heading', { name: 'Messages' })).toBeVisible();
+  // await page.getByRole('link', { name: 'Messages' }).click();
+  // await expect(page).toHaveURL('/account/messages/');
+  // await expect(page.getByRole('heading', { name: 'Messages' })).toBeVisible();
 
   await page.getByRole('link', { name: 'Addresses' }).click();
   await expect(page).toHaveURL('/account/addresses/');
   await expect(page.getByRole('heading', { name: 'Addresses' })).toBeVisible();
 
-  await page.getByRole('link', { name: 'Wish lists' }).click();
-  await expect(page).toHaveURL('/account/wishlists/');
-  await expect(page.getByRole('heading', { name: 'Wish lists' })).toBeVisible();
+  // await page.getByRole('link', { name: 'Wish lists' }).click();
+  // await expect(page).toHaveURL('/account/wishlists/');
+  // await expect(page.getByRole('heading', { name: 'Wish lists' })).toBeVisible();
 
-  await page.getByRole('link', { name: 'Recently viewed' }).click();
-  await expect(page).toHaveURL('/account/recently-viewed/');
-  await expect(page.getByRole('heading', { name: 'Recently viewed' })).toBeVisible();
+  // await page.getByRole('link', { name: 'Recently viewed' }).click();
+  // await expect(page).toHaveURL('/account/recently-viewed/');
+  // await expect(page.getByRole('heading', { name: 'Recently viewed' })).toBeVisible();
 
   await page.getByRole('link', { name: 'Account settings' }).click();
   await expect(page).toHaveURL('/account/settings/');
